requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
apipkg==1.5
Appium-Python-Client==1.0.2
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
blinker==1.4
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
et-xmlfile==1.0.1
execnet==1.7.1
Flask==1.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==1.1.1
mitmproxy==6.0.2
msgpack==1.0.2
openpyxl==3.0.5
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
protobuf==3.14.0
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.1
pytest-assume==2.4.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
tornado==6.1
urllib3==1.26.2
urwid==2.1.2
Werkzeug==1.0.1
wsproto==1.0.0
xmindparser==1.0.8
zstandard==0.14.1
